Tear down if we don't receive a pong in at least 3x our ping period
This doesn't align *exactly* with the protocol recommendations, but it
should do the job of noticing a server isn't sending us data.  The
spec suggests a typical keepalive is "a few minutes" and a client
should disconnect if it doesn't get a response "within a reasonable
amount of time after it has sent a PINGREQ."  For my applications, I
probably want to verify I'm getting useful publish messages for things
I'm subscribed to, but that logic is a bit too application specific
for me to understand how to fit it into this at the moment.