We don't need to generate client IDs for 3.1.1 either.
I was experiencing a misunderstanding of the spec combined with a bug
in mosquitto that made me believe this was required.

In particular, after all the things about requiring 1-23 character
client IDs to be supported, it says this:

  A Server MAY allow a Client to supply a ClientId that has a length
  of zero bytes, however if it does so the Server MUST treat this as a
  special case and assign a unique ClientId to that Client. It MUST
  then process the CONNECT packet as if the Client had provided that
  unique ClientId [MQTT-3.1.3-6].

...i.e., the server must support client IDs that are presented, but is
also welcome to accept zero-length client IDs and roll with it.

So I'll allow it all the way through.