waitForClient throws an exception now.
This is generally exceptional behavior and denying it makes it harder
to write reliable software (since I basically need to convert the
result back to an exception).

This also considers the final *state* of the client and throws an
exception when it's not good, as many types of disconnects were
detected properly, but not communicated properly.

Most code probably won't notice the difference.