Set mqtt-watch default timeout to 300s.
This is useful with how I use it.  This has the neat effect of
automatically resuming sessions if you drop a connection, at the cost
of a few extraneous sessions living about.

Of course, it would be kind to couple this with something to destroy
the session on ^C.  Maybe I should get to that as well.