When parsing pub segments, don't assume they don't encode 0 prop len
This is a slight misread of

    3.6.2.2.1 Property Length

    The length of the Properties in the PUBREL packet Variable Header
    encoded as a Variable Byte Integer. If the Remaining Length is
    less than 4 there is no Property Length and the value of 0 is
    used.

In my own encodings and all the ones I've encountered so far, the
property length is omitted whenever properties aren't supplied.  It
seems eqmx explicitly includes this property length.

This resulted in parsing the stream and having a leftover byte which
would be a misread and cause us to drop.  This happened to work
because if any property *were* present, the remaining length would be
> 4 as it'd have a key and value.  But in this exact case, the length
of 0 would cause my rl to be exactly 4.

Checking for an rl of > 3 is more exact and covers this case just
fine.  We do attempt to parse properties even when we know it's 0
which we could probably shortcut a bit, but that's an optimization.

Fixes #31